Simplified the ATerm library further. Removed blobs. Shared file i/o. Set the longs and ints of ATermTables and indexed sets to size_t. Removed ATdicts.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8463 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ 3rd-party/aterm/include/aterm/idmappings.h
@@ -30,7 +30,7 @@ typedef struct _IDMappings{
 	unsigned int threshold;
 } *IDMappings;
 
-IDMappings IMcreateIDMappings(float loadPercentage);
+IDMappings IMcreateIDMappings(unsigned int loadPercentage);
 
 int IMmakeIDMapping(IDMappings idMappings, void *key, unsigned int h, int value);
